Change in case of launched_at being NONE
A crash would occur at 140(in this version) if launched_at is None, so, for the time being, instead leave this server as it is, so as not make a decision for the user. This could be updated at a later point, by emailing such a server in the same way with the same message as other servers for example.